Move the background to create cover and stac item.
